// src/app/vacancy/interfaces/SourceKeys.enum.ts
enum SourceKeys {
    id = "id",
    busqueda_avature = "Busqueda avature",
    computrabajo = "Computrabajo",
    convocatoria = "Convocatoria",
    correo = "Correo",
    el_empleo = "El empleo",
    facebook = "Facebook",
    instagram = "Instagram",
    landing_page  = "Landing page",
    linkedIn = "LinkedIn",
    pandape = "Pandape",
    referido = "Referido",
    otros = "Otros",
    vacancy = "vacancy"

}

export default SourceKeys;