// src/app/vacancy/interfaces/SourceKeysOriginal.enum.ts
enum SourceKeysOriginal {
    id = "id",
    busqueda_avature = "busqueda_avature",
    computrabajo = "computrabajo",
    convocatoria = "convocatoria",
    correo = "correo",
    el_empleo = "el_empleo",
    facebook = "facebook",
    instagram = "instagram",
    landing_page  = "landing_page",
    linkedIn = "linkedIn",
    pandape = "pandape",
    referido = "referido",
    otros = "otros",
    vacancy = "vacancy"

}

export default SourceKeysOriginal;